fix: this is what happens when you don't go full-ass and just halfass everything.
The repos inspiring this setup don't use R2 as the primary backup destination; based on the configuration, it appears a local-ish minio is used, and then eventually it's pushed up to r2 from there? So because I'm skipping minio, there was no destination set up and the PVC won't finish provisioning, I think?